skript/requirements.txt
requests
python-dotenv
selenium
watchdog
2captcha-python
